DT-2004-11-12-0: o qmaster now uses very little memory for calculating subordinate queue suspension. o qmaster now ignores NULL job ids. o "" is now rejected as an invalid job id. o qconf now sets status to 1 for failed host group and cluster queue operations. o The lib path will now only be updated if the execd param, SET_LIB_PATH, is set. o The inheritance by jobs of environment of the shell which started the execd can   now be disabled with the INHERIT_ENV execd param. o qconf now correctly prints subordinate queue lists. o Subordinate queues are no properly suspended at qmaster startup. o Fixed broken build on Alpha 5 Issue number:  1289, 1291, 1295, 1300, 1324, 1335 Bugtraq number:   6194625, 6176181, 6174915, 6185211, 6193348, 6194713 Reviewed by:     AH, SG, AS, AA, EB

+++ source/clients/qhost/qhost.c
@@ -528,7 +528,7 @@ u_long32 show
 
          sge_dstring_clear(&resource_string);
 
-         switch (type) {
+         switch ((int)type) {
          case TYPE_HOST:   
          case TYPE_STR:   
          case TYPE_CSTR:
